// apps/admin-web/src/api/model/createNamespaceDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * I 54KB API
 * NestJS 企业级后端服务 API，基于 Prisma + PostgreSQL + Redis + Socket.io，提供用户认证、审计日志、幂等性控制等核心能力
 * OpenAPI spec version: 1.0
 */

export interface CreateNamespaceDto {
  /**
   * @minLength 1
   * @maxLength 50
   * @pattern ^[a-z][a-z0-9_]*$
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  displayName: string;
  /** @maxLength 500 */
  description?: string;
  isEnabled?: boolean;
}